+++ pyranha/type_system.cpp
@@ -32,12 +32,11 @@ see https://www.gnu.org/licenses/. */
 #include <boost/python/errors.hpp>
 #include <boost/python/object.hpp>
 #include <boost/python/stl_iterator.hpp>
+#include <boost/python/tuple.hpp>
 #include <cstddef>
 #include <functional>
 #include <string>
 #include <typeindex>
-#include <unordered_map>
-#include <unordered_set>
 #include <vector>
 
 #include "../src/config.hpp"
@@ -49,17 +48,19 @@ namespace pyranha
 
 namespace bp = boost::python;
 
-std::unordered_set<std::string> tg_names;
-std::unordered_map<std::type_index, bp::object> et_map;
-std::unordered_map<std::string, std::unordered_map<std::vector<std::type_index>, std::type_index, v_idx_hasher>>
-    gtg_map;
+// Map of exposed types.
+et_map_t et_map;
 
+// Map of registered template instances.
+ti_map_t ti_map;
+
+// Implementation of the methods of type_generator.
 bp::object type_generator::operator()() const
 {
     const auto it = et_map.find(m_t_idx);
     if (it == et_map.end()) {
         ::PyErr_SetString(PyExc_TypeError,
-                          ("the type '" + piranha::detail::demangle(m_t_idx) + "' has not been exposed").c_str());
+                          ("the type '" + piranha::detail::demangle(m_t_idx) + "' has not been registered").c_str());
         bp::throw_error_already_set();
     }
     return it->second;
@@ -70,6 +71,7 @@ std::string type_generator::repr() const
     return "Type generator for the C++ type '" + piranha::detail::demangle(m_t_idx) + "'";
 }
 
+// Implementation of the hasher for ti_map_t.
 std::size_t v_idx_hasher::operator()(const std::vector<std::type_index> &v) const
 {
     std::size_t retval = 0u;
@@ -80,43 +82,50 @@ std::size_t v_idx_hasher::operator()(const std::vector<std::type_index> &v) cons
     return retval;
 }
 
-type_generator generic_type_generator::operator()(bp::list l) const
-{
-    // We assume that this is created concurrently with the exposition of the gtg
-    // (and hence its registration on the C++ and Python sides).
-    piranha_assert(gtg_map.find(m_name) != gtg_map.end());
-    // Convert the list to a vector of type idx objects.
-    std::vector<std::type_index> v_t_idx;
-    bp::stl_input_iterator<type_generator> it(l), end;
-    for (; it != end; ++it) {
-        v_t_idx.push_back((*it).m_t_idx);
-    }
-    const auto it1 = gtg_map[m_name].find(v_t_idx);
-    if (it1 == gtg_map[m_name].end()) {
-        ::PyErr_SetString(PyExc_TypeError,
-                          ("the generic type generator '" + m_name + "' has not been instantiated with the type pack "
-                           + v_t_idx_to_str(v_t_idx))
-                              .c_str());
-        bp::throw_error_already_set();
-    }
-    return type_generator{it1->second};
-}
-
-std::string generic_type_generator::repr() const
-{
-    return "Type generator for the generic C++ type '" + m_orig_name + "'";
-}
-
-std::string v_t_idx_to_str(const std::vector<std::type_index> &v_t_idx)
+// Small utility to convert a vector of type indices to a string representation, only for error reporting purposes.
+static inline std::string v_t_idx_to_str(const std::vector<std::type_index> &v_t_idx)
 {
     std::string tv_name = "[";
     for (decltype(v_t_idx.size()) i = 0u; i < v_t_idx.size(); ++i) {
         tv_name += piranha::detail::demangle(v_t_idx[i]);
         if (i != v_t_idx.size() - 1u) {
-            tv_name += ",";
+            tv_name += ", ";
         }
     }
     tv_name += "]";
     return tv_name;
+}
+
+type_generator type_generator_template::getitem_t(bp::tuple t) const
+{
+    if (ti_map.find(m_name) == ti_map.end()) {
+        ::PyErr_SetString(PyExc_TypeError,
+                          ("no instance of the C++ class template '" + m_name + "' has been registered").c_str());
+        bp::throw_error_already_set();
+    }
+    // Convert the tuple of generators to a vector of type idx objects.
+    std::vector<std::type_index> v_t_idx;
+    bp::stl_input_iterator<type_generator> it(t), end;
+    for (; it != end; ++it) {
+        v_t_idx.push_back((*it).m_t_idx);
+    }
+    const auto it1 = ti_map[m_name].find(v_t_idx);
+    if (it1 == ti_map[m_name].end()) {
+        ::PyErr_SetString(PyExc_TypeError, ("no instance of the C++ class template '" + m_name
+                                            + "' has been registered with arguments " + v_t_idx_to_str(v_t_idx))
+                                               .c_str());
+        bp::throw_error_already_set();
+    }
+    return type_generator{it1->second};
+}
+
+type_generator type_generator_template::getitem_o(bp::object o) const
+{
+    return getitem_t(bp::make_tuple(o));
+}
+
+std::string type_generator_template::repr() const
+{
+    return "Type generator template for the C++ class template '" + m_name + "'";
 }
 }